% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fgammagpd.r
\name{fgammagpd}
\alias{fgammagpd}
\alias{lgammagpd}
\alias{nlgammagpd}
\alias{nlugammagpd}
\alias{proflugammagpd}
\title{MLE Fitting of Gamma Bulk and GPD Tail Extreme Value Mixture Model}
\usage{
fgammagpd(x, phiu = TRUE, useq = NULL, fixedu = FALSE, pvector = NULL,
  std.err = TRUE, method = "BFGS", control = list(maxit = 10000),
  finitelik = TRUE, ...)

lgammagpd(x, gshape = 1, gscale = 1, u = qgamma(0.9, gshape, 1/gscale),
  sigmau = sqrt(gshape) * gscale, xi = 0, phiu = TRUE, log = TRUE)

nlgammagpd(pvector, x, phiu = TRUE, finitelik = FALSE)

proflugammagpd(u, pvector, x, phiu = TRUE, method = "BFGS",
  control = list(maxit = 10000), finitelik = TRUE, ...)

nlugammagpd(pvector, u, x, phiu = TRUE, finitelik = FALSE)
}
\arguments{
\item{x}{vector of sample data}

\item{phiu}{probability of being above threshold \eqn{(0, 1)} or logical, see Details in
help for \code{\link[evmix:fnormgpd]{fnormgpd}}}

\item{useq}{vector of thresholds (or scalar) to be considered in profile likelihood or
\code{NULL} for no profile likelihood}

\item{fixedu}{logical, should threshold be fixed (at either scalar value in \code{useq},
or estimated from maximum of profile likelihood evaluated at
sequence of thresholds in \code{useq})}

\item{pvector}{vector of initial values of parameters or \code{NULL} for default
values, see below}

\item{std.err}{logical, should standard errors be calculated}

\item{method}{optimisation method (see \code{\link[stats:optim]{optim}})}

\item{control}{optimisation control list (see \code{\link[stats:optim]{optim}})}

\item{finitelik}{logical, should log-likelihood return finite value for invalid parameters}

\item{...}{optional inputs passed to \code{\link[stats:optim]{optim}}}

\item{gshape}{scalar gamma shape (positive)}

\item{gscale}{scalar gamma scale (positive)}

\item{u}{scalar threshold value}

\item{sigmau}{scalar scale parameter (positive)}

\item{xi}{scalar shape parameter}

\item{log}{logical, if \code{TRUE} then log-likelihood rather than likelihood is output}
}
\value{
Log-likelihood is given by \code{\link[evmix:fgammagpd]{lgammagpd}} and it's
  wrappers for negative log-likelihood from \code{\link[evmix:fgammagpd]{nlgammagpd}}
  and \code{\link[evmix:fgammagpd]{nlugammagpd}}. Profile likelihood for single
  threshold given by \code{\link[evmix:fgammagpd]{proflugammagpd}}. Fitting function
  \code{\link[evmix:fgammagpd]{fgammagpd}} returns a simple list with the
  following elements

\tabular{ll}{
 \code{call}:      \tab \code{optim} call\cr
 \code{x}:         \tab data vector \code{x}\cr
 \code{init}:      \tab \code{pvector}\cr
 \code{fixedu}:    \tab fixed threshold, logical\cr
 \code{useq}:      \tab threshold vector for profile likelihood or scalar for fixed threshold\cr
 \code{nllhuseq}:  \tab profile negative log-likelihood at each threshold in useq\cr
 \code{optim}:     \tab complete \code{optim} output\cr
 \code{mle}:       \tab vector of MLE of parameters\cr
 \code{cov}:       \tab variance-covariance matrix of MLE of parameters\cr
 \code{se}:        \tab vector of standard errors of MLE of parameters\cr
 \code{rate}:      \tab \code{phiu} to be consistent with \code{\link[evd:fpot]{evd}}\cr
 \code{nllh}:      \tab minimum negative log-likelihood\cr
 \code{n}:         \tab total sample size\cr
 \code{gshape}:    \tab MLE of gamma shape\cr
 \code{gscale}:    \tab MLE of gamma scale\cr
 \code{u}:         \tab threshold (fixed or MLE)\cr
 \code{sigmau}:    \tab MLE of GPD scale\cr
 \code{xi}:        \tab MLE of GPD shape\cr
 \code{phiu}:      \tab MLE of tail fraction (bulk model or parameterised approach)\cr
 \code{se.phiu}:   \tab standard error of MLE of tail fraction\cr
}
}
\description{
Maximum likelihood estimation for fitting the extreme value
mixture model with gamma for bulk distribution upto the threshold and conditional
GPD above threshold. With options for profile likelihood estimation for threshold and
fixed threshold approach.
}
\details{
The extreme value mixture model with gamma bulk and GPD tail is
fitted to the entire dataset using maximum likelihood estimation. The estimated
parameters, variance-covariance matrix and their standard errors are automatically
output.

See help for \code{\link[evmix:fnormgpd]{fnormgpd}} for details, type \code{help fnormgpd}.
Only the different features are outlined below for brevity.

The full parameter vector is
(\code{gshape}, \code{gscale}, \code{u}, \code{sigmau}, \code{xi}) if threshold is also estimated and
(\code{gshape}, \code{gscale}, \code{sigmau}, \code{xi}) for profile likelihood or fixed threshold approach.

Non-positive data are ignored as likelihood is infinite, except for \code{gshape=1}.
}
\note{
When \code{pvector=NULL} then the initial values are:
\itemize{
 \item approximation of MLE of gamma parameters assuming entire population is gamma; and
 \item threshold 90\% quantile (not relevant for profile likelihood for threshold or fixed threshold approaches);
 \item MLE of GPD parameters above threshold.
}
}
\section{Acknowledgments}{
 See Acknowledgments in
  \code{\link[evmix:fnormgpd]{fnormgpd}}, type \code{help fnormgpd}.
}
\examples{
\dontrun{
set.seed(1)
par(mfrow = c(2, 1))

x = rgamma(1000, shape = 2)
xx = seq(-0.1, 8, 0.01)
y = dgamma(xx, shape = 2)

# Bulk model based tail fraction
fit = fgammagpd(x)
hist(x, breaks = 100, freq = FALSE, xlim = c(-0.1, 8))
lines(xx, y)
with(fit, lines(xx, dgammagpd(xx, gshape, gscale, u, sigmau, xi), col="red"))
abline(v = fit$u, col = "red")

# Parameterised tail fraction
fit2 = fgammagpd(x, phiu = FALSE)
with(fit2, lines(xx, dgammagpd(xx, gshape, gscale, u, sigmau, xi, phiu), col="blue"))
abline(v = fit2$u, col = "blue")
legend("topright", c("True Density","Bulk Tail Fraction","Parameterised Tail Fraction"),
  col=c("black", "red", "blue"), lty = 1)

# Profile likelihood for initial value of threshold and fixed threshold approach
fitu = fgammagpd(x, useq = seq(1, 5, length = 20))
fitfix = fgammagpd(x, useq = seq(1, 5, length = 20), fixedu = TRUE)

hist(x, breaks = 100, freq = FALSE, xlim = c(-0.1, 8))
lines(xx, y)
with(fit, lines(xx, dgammagpd(xx, gshape, gscale, u, sigmau, xi), col="red"))
abline(v = fit$u, col = "red")
with(fitu, lines(xx, dgammagpd(xx, gshape, gscale, u, sigmau, xi), col="purple"))
abline(v = fitu$u, col = "purple")
with(fitfix, lines(xx, dgammagpd(xx, gshape, gscale, u, sigmau, xi), col="darkgreen"))
abline(v = fitfix$u, col = "darkgreen")
legend("topright", c("True Density","Default initial value (90\% quantile)",
 "Prof. lik. for initial value", "Prof. lik. for fixed threshold"),
 col=c("black", "red", "purple", "darkgreen"), lty = 1)
}
}
\author{
Yang Hu and Carl Scarrott \email{carl.scarrott@canterbury.ac.nz}
}
\references{
\url{http://www.math.canterbury.ac.nz/~c.scarrott/evmix}

\url{http://en.wikipedia.org/wiki/Gamma_distribution}

\url{http://en.wikipedia.org/wiki/Generalized_Pareto_distribution}

Scarrott, C.J. and MacDonald, A. (2012). A review of extreme value
threshold estimation and uncertainty quantification. REVSTAT - Statistical
Journal 10(1), 33-59. Available from \url{http://www.ine.pt/revstat/pdf/rs120102.pdf}

Hu, Y. (2013). Extreme value mixture modelling: An R package and simulation study.
MSc (Hons) thesis, University of Canterbury, New Zealand.
\url{http://ir.canterbury.ac.nz/simple-search?query=extreme&submit=Go}

Behrens, C.N., Lopes, H.F. and Gamerman, D. (2004). Bayesian analysis of extreme
events with threshold estimation. Statistical Modelling. 4(3), 227-244.
}
\seealso{
\code{\link[stats:GammaDist]{dgamma}},
 \code{\link[evmix:fgpd]{fgpd}} and \code{\link[evmix:gpd]{gpd}}

Other mgamma fmgamma
         gammagpd gammagpdcon fgammagpd fgammagpdcon normgpd fnormgpd
         mgammagpd mgammagpdcon fmgammagpd fmgammagpdcon: \code{\link{dgammagpdcon}},
  \code{\link{dgammagpdcon}}, \code{\link{dgammagpdcon}},
  \code{\link{dgammagpdcon}}, \code{\link{dgammagpdcon}},
  \code{\link{gammagpdcon}}, \code{\link{gammagpdcon}},
  \code{\link{gammagpdcon}}, \code{\link{gammagpdcon}},
  \code{\link{gammagpdcon}}, \code{\link{pgammagpdcon}},
  \code{\link{pgammagpdcon}}, \code{\link{pgammagpdcon}},
  \code{\link{pgammagpdcon}}, \code{\link{pgammagpdcon}},
  \code{\link{qgammagpdcon}}, \code{\link{qgammagpdcon}},
  \code{\link{qgammagpdcon}}, \code{\link{qgammagpdcon}},
  \code{\link{qgammagpdcon}}, \code{\link{rgammagpdcon}},
  \code{\link{rgammagpdcon}}, \code{\link{rgammagpdcon}},
  \code{\link{rgammagpdcon}}, \code{\link{rgammagpdcon}};
  \code{\link{dgammagpd}}, \code{\link{dgammagpd}},
  \code{\link{dgammagpd}}, \code{\link{dgammagpd}},
  \code{\link{dgammagpd}}, \code{\link{gammagpd}},
  \code{\link{gammagpd}}, \code{\link{gammagpd}},
  \code{\link{gammagpd}}, \code{\link{gammagpd}},
  \code{\link{pgammagpd}}, \code{\link{pgammagpd}},
  \code{\link{pgammagpd}}, \code{\link{pgammagpd}},
  \code{\link{pgammagpd}}, \code{\link{qgammagpd}},
  \code{\link{qgammagpd}}, \code{\link{qgammagpd}},
  \code{\link{qgammagpd}}, \code{\link{qgammagpd}},
  \code{\link{rgammagpd}}, \code{\link{rgammagpd}},
  \code{\link{rgammagpd}}, \code{\link{rgammagpd}},
  \code{\link{rgammagpd}}; \code{\link{dmgammagpdcon}},
  \code{\link{dmgammagpdcon}}, \code{\link{dmgammagpdcon}},
  \code{\link{dmgammagpdcon}}, \code{\link{dmgammagpdcon}},
  \code{\link{mgammagpdcon}}, \code{\link{mgammagpdcon}},
  \code{\link{mgammagpdcon}}, \code{\link{mgammagpdcon}},
  \code{\link{mgammagpdcon}}, \code{\link{pmgammagpdcon}},
  \code{\link{pmgammagpdcon}}, \code{\link{pmgammagpdcon}},
  \code{\link{pmgammagpdcon}}, \code{\link{pmgammagpdcon}},
  \code{\link{qmgammagpdcon}}, \code{\link{qmgammagpdcon}},
  \code{\link{qmgammagpdcon}}, \code{\link{qmgammagpdcon}},
  \code{\link{qmgammagpdcon}}, \code{\link{rmgammagpdcon}},
  \code{\link{rmgammagpdcon}}, \code{\link{rmgammagpdcon}},
  \code{\link{rmgammagpdcon}}, \code{\link{rmgammagpdcon}};
  \code{\link{dmgammagpd}}, \code{\link{dmgammagpd}},
  \code{\link{dmgammagpd}}, \code{\link{dmgammagpd}},
  \code{\link{dmgammagpd}}, \code{\link{mgammagpd}},
  \code{\link{mgammagpd}}, \code{\link{mgammagpd}},
  \code{\link{mgammagpd}}, \code{\link{mgammagpd}},
  \code{\link{pmgammagpd}}, \code{\link{pmgammagpd}},
  \code{\link{pmgammagpd}}, \code{\link{pmgammagpd}},
  \code{\link{pmgammagpd}}, \code{\link{qmgammagpd}},
  \code{\link{qmgammagpd}}, \code{\link{qmgammagpd}},
  \code{\link{qmgammagpd}}, \code{\link{qmgammagpd}},
  \code{\link{rmgammagpd}}, \code{\link{rmgammagpd}},
  \code{\link{rmgammagpd}}, \code{\link{rmgammagpd}},
  \code{\link{rmgammagpd}}; \code{\link{dmgamma}},
  \code{\link{dmgamma}}, \code{\link{dmgamma}},
  \code{\link{dmgamma}}, \code{\link{dmgamma}},
  \code{\link{mgamma}}, \code{\link{mgamma}},
  \code{\link{mgamma}}, \code{\link{mgamma}},
  \code{\link{mgamma}}, \code{\link{pmgamma}},
  \code{\link{pmgamma}}, \code{\link{pmgamma}},
  \code{\link{pmgamma}}, \code{\link{pmgamma}},
  \code{\link{qmgamma}}, \code{\link{qmgamma}},
  \code{\link{qmgamma}}, \code{\link{qmgamma}},
  \code{\link{qmgamma}}, \code{\link{rmgamma}},
  \code{\link{rmgamma}}, \code{\link{rmgamma}},
  \code{\link{rmgamma}}, \code{\link{rmgamma}};
  \code{\link{fgammagpdcon}}, \code{\link{fgammagpdcon}},
  \code{\link{fgammagpdcon}}, \code{\link{fgammagpdcon}},
  \code{\link{fgammagpdcon}}, \code{\link{lgammagpdcon}},
  \code{\link{lgammagpdcon}}, \code{\link{lgammagpdcon}},
  \code{\link{lgammagpdcon}}, \code{\link{lgammagpdcon}},
  \code{\link{nlgammagpdcon}}, \code{\link{nlgammagpdcon}},
  \code{\link{nlgammagpdcon}}, \code{\link{nlgammagpdcon}},
  \code{\link{nlgammagpdcon}},
  \code{\link{nlugammagpdcon}},
  \code{\link{nlugammagpdcon}},
  \code{\link{nlugammagpdcon}},
  \code{\link{nlugammagpdcon}},
  \code{\link{nlugammagpdcon}},
  \code{\link{proflugammagpdcon}},
  \code{\link{proflugammagpdcon}},
  \code{\link{proflugammagpdcon}},
  \code{\link{proflugammagpdcon}},
  \code{\link{proflugammagpdcon}};
  \code{\link{fmgammagpdcon}}, \code{\link{fmgammagpdcon}},
  \code{\link{fmgammagpdcon}}, \code{\link{fmgammagpdcon}},
  \code{\link{fmgammagpdcon}}, \code{\link{fmgammagpdcon}},
  \code{\link{fmgammagpdcon}}, \code{\link{lmgammagpdcon}},
  \code{\link{lmgammagpdcon}}, \code{\link{lmgammagpdcon}},
  \code{\link{lmgammagpdcon}}, \code{\link{lmgammagpdcon}},
  \code{\link{lmgammagpdcon}}, \code{\link{lmgammagpdcon}},
  \code{\link{nlEMmgammagpdcon}},
  \code{\link{nlEMmgammagpdcon}},
  \code{\link{nlEMmgammagpdcon}},
  \code{\link{nlEMmgammagpdcon}},
  \code{\link{nlEMmgammagpdcon}},
  \code{\link{nlEMmgammagpdcon}},
  \code{\link{nlEMmgammagpdcon}},
  \code{\link{nlmgammagpdcon}},
  \code{\link{nlmgammagpdcon}},
  \code{\link{nlmgammagpdcon}},
  \code{\link{nlmgammagpdcon}},
  \code{\link{nlmgammagpdcon}},
  \code{\link{nlmgammagpdcon}},
  \code{\link{nlmgammagpdcon}},
  \code{\link{nluEMmgammagpdcon}},
  \code{\link{nluEMmgammagpdcon}},
  \code{\link{nluEMmgammagpdcon}},
  \code{\link{nluEMmgammagpdcon}},
  \code{\link{nluEMmgammagpdcon}},
  \code{\link{nluEMmgammagpdcon}},
  \code{\link{nluEMmgammagpdcon}},
  \code{\link{nlumgammagpdcon}},
  \code{\link{nlumgammagpdcon}},
  \code{\link{nlumgammagpdcon}},
  \code{\link{nlumgammagpdcon}},
  \code{\link{nlumgammagpdcon}},
  \code{\link{nlumgammagpdcon}},
  \code{\link{nlumgammagpdcon}},
  \code{\link{proflumgammagpdcon}},
  \code{\link{proflumgammagpdcon}},
  \code{\link{proflumgammagpdcon}},
  \code{\link{proflumgammagpdcon}},
  \code{\link{proflumgammagpdcon}},
  \code{\link{proflumgammagpdcon}},
  \code{\link{proflumgammagpdcon}};
  \code{\link{fmgammagpd}}, \code{\link{fmgammagpd}},
  \code{\link{fmgammagpd}}, \code{\link{fmgammagpd}},
  \code{\link{fmgammagpd}}, \code{\link{fmgammagpd}},
  \code{\link{fmgammagpd}}, \code{\link{lmgammagpd}},
  \code{\link{lmgammagpd}}, \code{\link{lmgammagpd}},
  \code{\link{lmgammagpd}}, \code{\link{lmgammagpd}},
  \code{\link{lmgammagpd}}, \code{\link{lmgammagpd}},
  \code{\link{nlEMmgammagpd}}, \code{\link{nlEMmgammagpd}},
  \code{\link{nlEMmgammagpd}}, \code{\link{nlEMmgammagpd}},
  \code{\link{nlEMmgammagpd}}, \code{\link{nlEMmgammagpd}},
  \code{\link{nlEMmgammagpd}}, \code{\link{nlmgammagpd}},
  \code{\link{nlmgammagpd}}, \code{\link{nlmgammagpd}},
  \code{\link{nlmgammagpd}}, \code{\link{nlmgammagpd}},
  \code{\link{nlmgammagpd}}, \code{\link{nlmgammagpd}},
  \code{\link{nluEMmgammagpd}},
  \code{\link{nluEMmgammagpd}},
  \code{\link{nluEMmgammagpd}},
  \code{\link{nluEMmgammagpd}},
  \code{\link{nluEMmgammagpd}},
  \code{\link{nluEMmgammagpd}},
  \code{\link{nluEMmgammagpd}}, \code{\link{nlumgammagpd}},
  \code{\link{nlumgammagpd}}, \code{\link{nlumgammagpd}},
  \code{\link{nlumgammagpd}}, \code{\link{nlumgammagpd}},
  \code{\link{nlumgammagpd}}, \code{\link{nlumgammagpd}},
  \code{\link{proflumgammagpd}},
  \code{\link{proflumgammagpd}},
  \code{\link{proflumgammagpd}},
  \code{\link{proflumgammagpd}},
  \code{\link{proflumgammagpd}},
  \code{\link{proflumgammagpd}},
  \code{\link{proflumgammagpd}}; \code{\link{fmgamma}},
  \code{\link{fmgamma}}, \code{\link{fmgamma}},
  \code{\link{fmgamma}}, \code{\link{lmgamma}},
  \code{\link{lmgamma}}, \code{\link{lmgamma}},
  \code{\link{lmgamma}}, \code{\link{nlEMmgamma}},
  \code{\link{nlEMmgamma}}, \code{\link{nlEMmgamma}},
  \code{\link{nlEMmgamma}}, \code{\link{nlmgamma}},
  \code{\link{nlmgamma}}, \code{\link{nlmgamma}},
  \code{\link{nlmgamma}}
}

